creating terms.css ,terms.js, auth.php, db_connect.php, terms&conditions.php
terms.js is the java script file for the terms&conditions.php file, the terms.css is the styling for the php. auth.php is the page responsible for the register/login for the user and the admin, also the db_connect.php is the one used for connecting the database with the code.